Refactor TargetProfile, AttackStep, and AttackChain classes for enhanced attack modeling
