+++ hexstrike_server.py
@@ -86,105 +86,10 @@ API_HOST = os.environ.get('HEXSTRIKE_HOST', '127.0.0.1')
 # INTELLIGENT DECISION ENGINE (v6.0 ENHANCEMENT)
 # ============================================================================
 from shared.target_types import TargetType, TechnologyStack
+from shared.target_profile import TargetProfile
+from shared.attack_step import AttackStep
+from shared.attack_chain import AttackChain
 
-@dataclass
-class TargetProfile:
-    """Comprehensive target analysis profile for intelligent decision making"""
-    target: Any
-    target_type: TargetType = TargetType.UNKNOWN
-    ip_addresses: List[str] = field(default_factory=list)
-    open_ports: List[int] = field(default_factory=list)
-    services: Dict[int, str] = field(default_factory=dict)
-    technologies: List[TechnologyStack] = field(default_factory=list)
-    cms_type: Optional[str] = None
-    cloud_provider: Optional[str] = None
-    security_headers: Dict[str, str] = field(default_factory=dict)
-    ssl_info: Dict[str, Any] = field(default_factory=dict)
-    subdomains: List[str] = field(default_factory=list)
-    endpoints: List[str] = field(default_factory=list)
-    attack_surface_score: float = 0.0
-    risk_level: str = "unknown"
-    confidence_score: float = 0.0
-
-    def to_dict(self) -> Dict[str, Any]:
-        """Convert TargetProfile to dictionary for JSON serialization"""
-        return {
-            "target": self.target,
-            "target_type": self.target_type.value,
-            "ip_addresses": self.ip_addresses,
-            "open_ports": self.open_ports,
-            "services": self.services,
-            "technologies": [tech.value for tech in self.technologies],
-            "cms_type": self.cms_type,
-            "cloud_provider": self.cloud_provider,
-            "security_headers": self.security_headers,
-            "ssl_info": self.ssl_info,
-            "subdomains": self.subdomains,
-            "endpoints": self.endpoints,
-            "attack_surface_score": self.attack_surface_score,
-            "risk_level": self.risk_level,
-            "confidence_score": self.confidence_score
-        }
-
-@dataclass
-class AttackStep:
-    """Individual step in an attack chain"""
-    tool: str
-    parameters: Dict[str, Any]
-    expected_outcome: str
-    success_probability: float
-    execution_time_estimate: int  # seconds
-    dependencies: List[str] = field(default_factory=list)
-
-class AttackChain:
-    """Represents a sequence of attacks for maximum impact"""
-    def __init__(self, target_profile: TargetProfile):
-        self.target_profile = target_profile
-        self.steps: List[AttackStep] = []
-        self.success_probability: float = 0.0
-        self.estimated_time: int = 0
-        self.required_tools: Set[str] = set()
-        self.risk_level: str = "unknown"
-
-    def add_step(self, step: AttackStep):
-        """Add a step to the attack chain"""
-        self.steps.append(step)
-        self.required_tools.add(step.tool)
-        self.estimated_time += step.execution_time_estimate
-
-    def calculate_success_probability(self):
-        """Calculate overall success probability of the attack chain"""
-        if not self.steps:
-            self.success_probability = 0.0
-            return
-
-        # Use compound probability for sequential steps
-        prob = 1.0
-        for step in self.steps:
-            prob *= step.success_probability
-
-        self.success_probability = prob
-
-    def to_dict(self) -> Dict[str, Any]:
-        """Convert AttackChain to dictionary"""
-        return {
-            "target": self.target_profile.target,
-            "steps": [
-                {
-                    "tool": step.tool,
-                    "parameters": step.parameters,
-                    "expected_outcome": step.expected_outcome,
-                    "success_probability": step.success_probability,
-                    "execution_time_estimate": step.execution_time_estimate,
-                    "dependencies": step.dependencies
-                }
-                for step in self.steps
-            ],
-            "success_probability": self.success_probability,
-            "estimated_time": self.estimated_time,
-            "required_tools": list(self.required_tools),
-            "risk_level": self.risk_level
-        }
 
 class IntelligentDecisionEngine:
     """AI-powered tool selection and parameter optimization engine"""
