+++ shared/attack_chain.py
@@ -1,0 +1,54 @@
+
+from typing import List, Set, Dict, Any
+from shared.target_profile import TargetProfile
+from shared.attack_step import AttackStep
+
+class AttackChain:
+    """Represents a sequence of attacks for maximum impact"""
+    def __init__(self, target_profile: TargetProfile):
+        self.target_profile = target_profile
+        self.steps: List[AttackStep] = []
+        self.success_probability: float = 0.0
+        self.estimated_time: int = 0
+        self.required_tools: Set[str] = set()
+        self.risk_level: str = "unknown"
+
+    def add_step(self, step: AttackStep):
+        """Add a step to the attack chain"""
+        self.steps.append(step)
+        self.required_tools.add(step.tool)
+        self.estimated_time += step.execution_time_estimate
+
+    def calculate_success_probability(self):
+        """Calculate overall success probability of the attack chain"""
+        if not self.steps:
+            self.success_probability = 0.0
+            return
+
+        # Use compound probability for sequential steps
+        prob = 1.0
+        for step in self.steps:
+            prob *= step.success_probability
+
+        self.success_probability = prob
+
+    def to_dict(self) -> Dict[str, Any]:
+        """Convert AttackChain to dictionary"""
+        return {
+            "target": self.target_profile.target,
+            "steps": [
+                {
+                    "tool": step.tool,
+                    "parameters": step.parameters,
+                    "expected_outcome": step.expected_outcome,
+                    "success_probability": step.success_probability,
+                    "execution_time_estimate": step.execution_time_estimate,
+                    "dependencies": step.dependencies
+                }
+                for step in self.steps
+            ],
+            "success_probability": self.success_probability,
+            "estimated_time": self.estimated_time,
+            "required_tools": list(self.required_tools),
+            "risk_level": self.risk_level
+        }
