+++ shared/attack_step.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass, field
+from typing import Any, Dict, List
+
+@dataclass
+class AttackStep:
+    """Individual step in an attack chain"""
+    tool: str
+    parameters: Dict[str, Any]
+    expected_outcome: str
+    success_probability: float
+    execution_time_estimate: int  # seconds
+    dependencies: List[str] = field(default_factory=list)
